+++ docs/release-notes.md
@@ -49,7 +49,12 @@ and
 #### Bug Fixes
 
 - `@ForAll` annotation could not be replaced by self-made annotation. Now it can. 
-- Annotations in type parameters of bounds of type variables are now correctly recognized.
+
+- Annotations in type parameters of bounds of type variables 
+  are now correctly recognized.
+
+- Annotations in type parameters of bounds of wildcards 
+  are now correctly recognized.
 
 ## 1.2.0